requirements.txt
google-api-python-client 
google-auth 
google-auth-httplib2 
google-auth-oauthlib
pixoo
viam-sdk==0.50.0

typing-extensions
